[Refactor] Move connector bootstrap into ModuleBootstrap
