+++ be/src/connector/cache_stats/AGENTS.md
@@ -11,5 +11,5 @@ Cache-stats connector implementation above connector contracts with storage acce
 - Allowed internal include prefixes: `connector/cache_stats/`, `connector/connector.h`, `connector/data_source.h`, `connector/data_source_provider.h`, `storage/`, `exprs/`, `runtime/`, `column/`, `types/`, `common/`, `base/`, `gutil/`, `gen_cpp/`
 - Allowed target deps: `ConnectorPrimitive`, `Storage`, `Expr`, `Runtime`, `ChunkCore`, `ColumnCore`, `Types`, `Common`, `Base`, `Gutil`, `StarRocksGen`
 - Core tests: `connector_cache_stats_test`
-- Remediation: Keep ConnectorCacheStats limited to the cache-stats connector and scanner implementation over Storage; move registration into ConnectorBootstrap and avoid pulling Connector, service, cache, or full Exec code into the connector library.
+- Remediation: Keep ConnectorCacheStats limited to the cache-stats connector and scanner implementation over Storage; move registration into ModuleBootstrap and avoid pulling Connector, service, cache, or full Exec code into the connector library.
 <!-- END GENERATED: BE MODULE HARNESSES -->
