+++ be/src/module/connector_bootstrap.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "connector/connector_bootstrap.h"
+#include "module/connector_bootstrap.h"
 
 #include <memory>
 #include <string>
