+++ be/src/service/CMakeLists.txt
@@ -36,7 +36,7 @@ ADD_BE_LIB(Service
 
 target_link_libraries(Service PRIVATE Script Runtime Orchestration Cache AgentServer ComputeEnv ExecRuntime Platform ColumnCore)
 target_link_libraries(ServiceBE PUBLIC Service)
-target_link_libraries(ServiceBE PRIVATE AgentServer ConnectorBootstrap DataWorkflows Orchestration Storage)
+target_link_libraries(ServiceBE PRIVATE AgentServer ModuleBootstrap DataWorkflows Orchestration Storage)
 
 # only build starrocks_be when TEST is off
 if ((NOT ${MAKE_TEST} STREQUAL "ON") AND (NOT BUILD_FORMAT_LIB))
