+++ be/src/service/service_be/starrocks_be.cpp
@@ -40,12 +40,12 @@
 #include "common/thread/priority_thread_pool.hpp"
 #include "compute_env/compute_env.h"
 #include "compute_env/staros/staros_worker_runtime.h"
-#include "connector/connector_bootstrap.h"
 #include "data_workflows/data_workflows_env.h"
 #include "exec/exec_env.h"
 #include "exec/pipeline/driver_executor_factory.h"
 #include "exec/pipeline/driver_queue_factory.h"
 #include "exec/pipeline/primitives/pipeline_metrics.h"
+#include "module/connector_bootstrap.h"
 #include "orchestration/orchestration_env.h"
 #include "platform/platform_env.h"
 #include "platform/store_path.h"
