+++ be/test/CMakeLists.txt
@@ -231,6 +231,7 @@ add_subdirectory(exprs)
 add_subdirectory(io)
 add_subdirectory(fs)
 add_subdirectory(fs_ext)
+add_subdirectory(module)
 add_subdirectory(platform)
 add_subdirectory(geo)
 add_subdirectory(cache)
