+++ be/test/module/CMakeLists.txt
@@ -1,0 +1,28 @@
+# Copyright 2021-present StarRocks, Inc. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+set(MODULE_BOOTSTRAP_TEST_LINK_LIBS
+    ModuleBootstrap
+    ${STARROCKS_DEPENDENCIES}
+    ${STARROCKS_STATIC_LINK_LIBS}
+    ${WL_LINK_DYNAMIC} ${STARROCKS_SYSTEM_LINK_LIBS} ${WRAP_LINKER_FLAGS}
+    ${WL_START_GROUP}
+    gmock
+    gtest
+    ${WL_END_GROUP}
+)
+
+add_executable(module_bootstrap_test module_bootstrap_test.cpp)
+target_link_libraries(module_bootstrap_test ${MODULE_BOOTSTRAP_TEST_LINK_LIBS} gtest_main)
+set_target_properties(module_bootstrap_test PROPERTIES COMPILE_FLAGS "-fno-access-control")
