+++ be/test/testutil/init_test_env.h
@@ -36,7 +36,6 @@
 #include "common/system/mem_info.h"
 #include "common/thread/priority_thread_pool.hpp"
 #include "compute_env/compute_env.h"
-#include "connector/connector_bootstrap.h"
 #include "data_workflows/data_workflows_env.h"
 #include "exec/exec_env.h"
 #include "exec/pipeline/driver_executor_factory.h"
@@ -44,6 +43,7 @@
 #include "exec/pipeline/primitives/pipeline_metrics.h"
 #include "exec/pipeline/query_context.h"
 #include "gtest/gtest.h"
+#include "module/connector_bootstrap.h"
 #include "module/fs_provider_bootstrap.h"
 #include "orchestration/orchestration_env.h"
 #include "platform/platform_env.h"
